// code-drop-front/types/languages.ts
export type SupportedLanguages =
  | "bash"
  | "c"
  | "clojure"
  | "cpp"
  | "csharp"
  | "dart"
  | "elixir"
  | "elm"
  | "erlang"
  | "fsharp"
  | "graphql"
  | "go"
  | "groovy"
  | "haskell"
  | "html"
  | "java"
  | "javascript"
  | "jsx"
  | "julia"
  | "kotlin"
  | "lisp"
  | "makefile"
  | "matlab"
  | "objectivec"
  | "ocaml"
  | "php"
  | "python"
  | "r"
  | "ruby"
  | "rust"
  | "scala"
  | "sql"
  | "swift"
  | "tsx"
  | "typescript";